+++ index.js
@@ -75,7 +75,6 @@ export { default as Modal } from './components/modal/Modal';
 export { default as FooterModal } from './components/modal/Footer';
 export { default as Price } from './components/price/Price';
 export { default as Progress } from './components/progress/Progress';
-export { default as Text } from './components/text/Text';
 export { default as Legend } from './components/text/Legend';
 export { default as ErrorZone } from './components/text/ErrorZone';
 export { default as LanguageSection } from './containers/general/LanguageSection';
@@ -209,7 +208,6 @@ export { default as LayoutsSection } from './containers/layouts/LayoutsSection';
 export { default as ToolbarsSection } from './containers/layouts/ToolbarsSection';
 export { default as ViewLayoutRadios } from './containers/layouts/ViewLayoutRadios';
 export { default as ComposerModeRadios } from './containers/layouts/ComposerModeRadios';
-export { default as ThemeSection } from './containers/themes/ThemeSection';
 export { default as ThemesSection } from './containers/themes/ThemesSection';
 export { default as ThemeCard } from './containers/themes/ThemeCard';
 export { default as ThemeCards } from './containers/themes/ThemeCards';
